ex3_float.py added

+++ ex3_float.py
@@ -4,7 +4,7 @@ print("I will now count my chickens:")
 # Evaluated according to PEMDAS
     # 30/6 ---> 5
     # 25 + 5 = 30
-print("Hens", 25 + 30 /6)
+print("Hens", float(25 + 30 /6))
 
 
 # -25 * 3 evaluated first
@@ -13,7 +13,7 @@ print("Hens", 25 + 30 /6)
             # % is modulo operation
         # 100 -3 
         # 97
-print("Roosters", 100 -25 * 3 % 4)
+print("Roosters", float(100 -25 * 3 % 4))
 
 print("Now I will count the eggs:")
 
@@ -21,12 +21,12 @@ print("Now I will count the eggs:")
     # 3+2 + 1 - 5 + 4 % 2 - 0.25 +6
     # 4 % 2 modulo evaluated next
         #  3+2 + 1 - 5 + 0 - 0.25 + 6
-print(3+2 + 1 - 5 + 4 % 2 - 1 /4+6)
+print(float(3+2 + 1 - 5 + 4 % 2 - 1 /4+6))
 
 print("Is it true that 3 + 2 < 5 - 7?")
 
 # Evaluates 5 < -2 --> output is false
-print(3 + 2 < 5 - 7)
+print(float(3 + 2 < 5 - 7))
 
 print("What is 3 + 2", 3+2)
 print("What is 5-7?", 5-7)
